Website Update Commits
Introduced a Navigation Bar, which I am still very much working on.  Changed colors around a little bit. Added a 'Contact Me' button ,which opens to an email, at the bottom of the page. Added Icons to represent the "Skills" section and floated them to the left and right of their descriptions to increase readability. Started 'blog.html' which I will be working on, the goal is to start a blog to capture the progress I make and learn while I do it.